PMM-10733-DB-version-selection-simplification: move db version field to advanced options
