+++ public/app/percona/dbaas/components/DBCluster/AddDBClusterModal/DBClusterBasicOptions/DBClusterBasicOptions.tsx
@@ -1,25 +1,19 @@
 import { TextInputField, validators } from '@percona/platform-core';
-import React, { FC, useCallback, useState, useMemo } from 'react';
+import React, { FC, useCallback, useState } from 'react';
 import { Field } from 'react-final-form';
 
-import { SelectableValue } from '@grafana/data';
 import { Messages } from 'app/percona/dbaas/DBaaS.messages';
-import {
-  SelectFieldAdapter,
-  AsyncSelectFieldAdapter,
-} from 'app/percona/shared/components/Form/FieldAdapters/FieldAdapters';
+import { SelectFieldAdapter } from 'app/percona/shared/components/Form/FieldAdapters/FieldAdapters';
 import { Databases } from 'app/percona/shared/core';
 
 import { Kubernetes, Operator } from '../../../Kubernetes/Kubernetes.types';
 import { getDatabaseOptionFromOperator } from '../../../Kubernetes/Kubernetes.utils';
 import { KubernetesOperatorStatus } from '../../../Kubernetes/OperatorStatusItem/KubernetesOperatorStatus/KubernetesOperatorStatus.types';
 import { DATABASE_OPTIONS } from '../../DBCluster.constants';
-import { isOptionEmpty } from '../../DBCluster.utils';
 import { AddDBClusterFields } from '../AddDBClusterModal.types';
 import { DBClusterTopology } from '../DBClusterAdvancedOptions/DBClusterAdvancedOptions.types';
 
 import { CLUSTER_NAME_MAX_LENGTH } from './DBClusterBasicOptions.constants';
-import { useDatabaseVersions } from './DBClusterBasicOptions.hooks';
 import { DatabaseOption, DBClusterBasicOptionsProps, Operators } from './DBClusterBasicOptions.types';
 import { getKubernetesOptions, kubernetesClusterNameValidator, optionRequired } from './DBClusterBasicOptions.utils';
 
@@ -40,9 +34,7 @@ const getAvailableDatabaseOptions = (kubernetesCluster: Kubernetes): DatabaseOpt
 export const DBClusterBasicOptions: FC<DBClusterBasicOptionsProps> = ({ kubernetes, form }) => {
   const { required, maxLength } = validators;
   const { change } = form;
-  const { kubernetesCluster, databaseType } = form.getState().values;
-  const [databaseVersions, setDatabaseVersions] = useState<SelectableValue[]>([]);
-  const [loadingDatabaseVersions, setLoadingDatabaseVersions] = useState(false);
+  const { kubernetesCluster } = form.getState().values;
   const onChangeDatabase = useCallback((databaseType) => {
     if (databaseType.value !== Databases.mysql) {
       change(AddDBClusterFields.topology, DBClusterTopology.cluster);
@@ -57,8 +49,7 @@ export const DBClusterBasicOptions: FC<DBClusterBasicOptionsProps> = ({ kubernet
 
   const [databaseOptions, setDatabaseOptions] = useState(() => {
     if (kubernetesCluster) {
-      const availableDatabaseOptions = getAvailableDatabaseOptions(kubernetesCluster);
-      return availableDatabaseOptions;
+      return getAvailableDatabaseOptions(kubernetesCluster);
     }
     return DATABASE_OPTIONS;
   });
@@ -79,10 +70,6 @@ export const DBClusterBasicOptions: FC<DBClusterBasicOptionsProps> = ({ kubernet
     change(AddDBClusterFields.kubernetesCluster, selectedKubernetes);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
-
-  const isDatabaseVersionDisabled = useMemo(() => isOptionEmpty(databaseType), [databaseType]);
-
-  useDatabaseVersions(form, databaseType, kubernetesCluster, setLoadingDatabaseVersions, setDatabaseVersions);
 
   return (
     <>
@@ -105,16 +92,6 @@ export const DBClusterBasicOptions: FC<DBClusterBasicOptionsProps> = ({ kubernet
         component={SelectFieldAdapter}
         validate={optionRequired}
         onChange={onChangeDatabase}
-      />
-      <Field
-        disabled={isDatabaseVersionDisabled}
-        dataTestId="dbcluster-database-version-field"
-        name={AddDBClusterFields.databaseVersion}
-        label={Messages.dbcluster.addModal.fields.databaseVersion}
-        component={AsyncSelectFieldAdapter}
-        loading={loadingDatabaseVersions}
-        options={databaseVersions}
-        validate={optionRequired}
       />
       <TextInputField
         name={AddDBClusterFields.name}
